+++ src/app/api/deleteTask/route.ts
@@ -1,14 +1,25 @@
 import { NextResponse } from "next/server";
 import connectToDb from "@/dbconfig/connectToDb";
 import Task from "@/models/taskModel";
+import { z } from "zod";
+
+const bodySchema = z.object({
+  _id: z.string(),
+});
 
 export async function POST(request: Request) {
   try {
     await connectToDb();
+
     const req = await request.json();
+
+    if (!bodySchema.safeParse(req).success) {
+      return NextResponse.json({ success: false, error: "Invalid request" });
+    }
+
     const { _id } = req;
 
-    const task = await Task.findOneAndDelete({ _id: _id });
+    await Task.findOneAndDelete({ _id: _id });
 
     return NextResponse.json({ success: true });
   } catch (error) {
